refactor(translation): reorganize library into domain modules
Replace flat translation helpers with cohesive domain modules for context,
concordance, generation, jobs, CAT, sandbox, and file memory while updating
callers to import the new concrete module paths.
