+++ apps/hyperlocalise-web/src/lib/agents/github/github-repository-automation-hl-check.ts
@@ -1,9 +1,5 @@
 import type { HlCheckReport } from "@/lib/providers/provider-job-qa/hl-check-types";
-import {
-  prepareSandbox,
-  runSandboxCommand,
-  writeFileToSandbox,
-} from "@/lib/translation/sandbox-translation";
+import { prepareSandbox, runSandboxCommand, writeFileToSandbox } from "@/lib/translation/sandbox";
 import { isErr } from "@/lib/primitives/result/results";
 import { safeJsonParse } from "@/lib/primitives/safeJsonParse/safeJsonParse";
 import { shellQuote } from "@/lib/primitives/shell-quote/shell-quote";
