--- apps/hyperlocalise-web/src/lib/translation/translation-memory-search-query.ts
@@ -1,12 +1,0 @@
-const maxContextSearchTerms = 50;
-
-export function buildTranslationMemoryTsQuery(input: string): string {
-  return input
-    .replace(/[&|!():*<>'"-]/g, " ")
-    .trim()
-    .split(/\s+/)
-    .filter(Boolean)
-    .slice(0, maxContextSearchTerms)
-    .map((word) => `${word}:*`)
-    .join(" & ");
-}
